🧑‍💻 refactor: tee form content
